barebox: reimplement installation method
Depending on features like habv4, IMGBASE can be list with multiple
extensions.  There was already basic, but broken support for iterating
over this list (the ELF file code expected only a single value).

It can be also useful to assign a certain name to the deployed image;
this was implemented by supporting a ':' delimeter like in

| IMGBASE = "<src-like-built-by-barebox>:<target-to-be-deployed>"

Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>